Experimenting with different bandwidth signal for denoising module + SSL with validation loop on classifier (MLP) 😊
Q: How would the denoiser perform when trained on different bandwidth of the same signal (pulse shape with different sps)?

E: Generate 7000 training examples of the QPSK signal pulse shaped at 5 different values from 2 to 32 at different SNR values between -20 to 20. Investigate the response of the learned model and the output of the model against the ground truth

Observation: The frequency response of the trained model is able to closely reconstruct the noisy signal of any bandwidth however it lets through high frequency signals outside the band of interest (This is the result for a preliminary training)